eco-logic-backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-generativeai
Pillow>=10.0.0
tavily-python
Jinja2>=3.1.2
python-multipart>=0.0.6
# firebase-admin (no longer needed, replaced with MongoDB)
typing-extensions>=4.8.0
pydantic>=2.4.2
httpx>=0.24.1
# MongoDB requirements
pymongo>=4.5.0
dnspython>=2.4.0
# uuid is part of Python's standard library, no need to install
# fastapi-cors is not needed as CORS is handled by FastAPI directly
